+++ src/main/java/core/base/syntax/model/Dog.java
@@ -1,0 +1,6 @@
+package CoreBaseSyntax.Model;
+
+public class Dog {
+    public Dog() {
+    }
+}
