+++ src/test/java/core/base/syntax/HelloWorldTest.java
@@ -1,4 +1,4 @@
-package core.basesyntax;
+package CoreBaseSyntax;
 
 /**
  * Feel free to remove this class and create your own.
